+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed tp06 field from Graphcast operational model
+- Removed static fields from Graphcast model input / outputs
 - Moved StormCast and DLESyM checkpoints to Huggingface
 
 ### Deprecated